requirements.txt
fastapi
uvicorn
python-dotenv
ldap3